types: refactor dir structure
