+++ packages/types/tests/store.creation.inference.test.ts
@@ -1,4 +1,4 @@
-import { ActionsType, StoreConfig } from "..";
+import { ActionsType, StoreConfig } from "../src";
 import {
   CompositeStoreState,
   SimpleStoreState,
